% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSSRMatrix.R
\name{calculateSSRMatrix}
\alias{calculateSSRMatrix}
\title{A function that calculate and returns the ssr (sum of squared residual) of all possible combinations of segments in a matrix form.
The index of row and column represents the starting and ending point of the segment.
For example, value at entry [1,100] of the matrix is the SSR of the segment
starting at data 1 and ending at data 100.}
\usage{
calculateSSRMatrix(data)
}
\arguments{
\item{data}{The input data}
}
\value{
SSRMatrix
}
\description{
A function that calculate and returns the ssr (sum of squared residual) of all possible combinations of segments in a matrix form.
The index of row and column represents the starting and ending point of the segment.
For example, value at entry [1,100] of the matrix is the SSR of the segment
starting at data 1 and ending at data 100.
}

