% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBP.R
\name{findBP}
\alias{findBP}
\title{A function that finds the breakpoints (BP) with the minimum ssr, given the number of desired BP by searching through the given ssr matrix}
\usage{
findBP(SSRMatrix, noOfBP, l, index, dataSize)
}
\arguments{
\item{SSRMatrix}{The ssrmatrix calculate by calculateSSRMatrix}

\item{noOfBP}{The desired number of breakpoints}

\item{l}{The minimum distance between each breakpoints}

\item{index}{The index that the current loop is on,}

\item{dataSize}{The data size for the remaining loop}
}
\value{
The set of breakpoints with the minimum ssr
}
\description{
A function that finds the breakpoints (BP) with the minimum ssr, given the number of desired BP by searching through the given ssr matrix
}

