% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findoneBP.R
\name{findoneBP}
\alias{findoneBP}
\title{A function that is used to find one breakpoint only, without creating the SSR matrix.}
\usage{
findoneBP(data, l)
}
\arguments{
\item{data}{The dataset that wants to be approximated}

\item{l}{The minimum size of a segment}
}
\value{
The breakpoint that gives the minimum ssr
}
\description{
A function that is used to find one breakpoint only, without creating the SSR matrix.
}

