% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getequations.R
\name{getequations}
\alias{getequations}
\title{A function that gets the equations with the given the breakpoints.}
\usage{
getequations(data, BPs)
}
\arguments{
\item{data}{The data to be approximated}

\item{BPs}{The list of breakpoints}
}
\value{
The coefficients of the equations, fitted values, and residuals.
}
\description{
A function that gets the equations with the given the breakpoints.
}

