% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pwl.r
\name{plot.pwl}
\alias{plot.pwl}
\title{A function that plots the piecewise equations along with the dataset}
\usage{
\method{plot}{pwl}(pwl, data)
}
\arguments{
\item{pwl}{The approximated piecewise linear equation.}

\item{data}{The data that the user want to plot with}
}
\value{
The plot of the data with the provided pwl
}
\description{
A function that plots the piecewise equations along with the dataset
}

