% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicts.pwl.r
\name{predicts.pwl}
\alias{predicts.pwl}
\title{A function that uses the piecewise equations to predict new values.}
\usage{
predicts.pwl(pwl, newData)
}
\arguments{
\item{pwl}{The piecewise linear equation}

\item{newData}{The new data [x, y] that user wants to predict values with the provided pwl}
}
\value{
The predicted values, y', using the provided pwl equation
}
\description{
A function that uses the piecewise equations to predict new values.
}

